fix: Convert the API priorities into a consumable priority (#115)
# Pull Request

Closes #112 

## Short description

We recently decided that our priorities should be based on what the user
expectation is, rather than what the API actually is. So essentially
priority of 1 is actually just a p4. What we forgot to do was during the
task mapping, we weren't updating the priorities in the task data going
out of the MCP

## PR Checklist

Feel free to leave unchecked or remove the lines that are not
applicable.

-   [ ] Added tests for bugs / new features
-   [ ] Updated docs (README, etc.)
- [ ] New tools added to `getMcpServer` AND exported in `src/index.ts`.

<!--
_Note:_ versioning is handled by
[release-please](https://github.com/googleapis/release-please) action,
based on the PR title.
-->