fix: dayCount and overdue only fixes (#128)
## Short description

There are two fixes here:
- If you specify a date, but don't provide anything for the daysCount,
it would default to 1, and then we would add another 1 onto that. So
effectively, if I gave the date of 2025-10-02, it would be anything from
2025-10-02 to 2025-10-03, which is definitely not what the user will
have requested.
- With a recent change (#108) we were defaulting to including overdue
items when the user put `today` as the startDate, but we weren't
providing a way to either not include overdue tasks, or a way to fetch
overdue tasks only.

## PR Checklist

Feel free to leave unchecked or remove the lines that are not
applicable.

-   [ ] Added tests for bugs / new features
-   [ ] Updated docs (README, etc.)
- [ ] New tools added to `getMcpServer` AND exported in `src/index.ts`.

<!--
_Note:_ versioning is handled by
[release-please](https://github.com/googleapis/release-please) action,
based on the PR title.
-->